Added some documentation
