+++ Sources/Aurora/Aurora Networking/Aurora.Networking.swift
@@ -101,62 +101,17 @@ extension Aurora {
         return HTTPSCertificate.publicKeyHash
     }
     
-    /// <#Description#>
-    /// - Parameter fromURL: <#fromURL description#>
-    /// - Returns: <#description#>
-    @available(*, deprecated)
-    func networkFetch(fromURL: URL) -> Data {
-        var waiting = true
-        var returnData: Data = Data.init()
-        
-        URLSession.shared.dataTask(with: fromURL) { dataTaskData, _, _ in
-            if let dataTaskData = dataTaskData {
-                returnData = dataTaskData
-            }
-            
-            waiting = false
-        }
-        .resume()
-        
-        while waiting { }
-        
-        return returnData
-    }
-    
-    /// networkRequest (blocking)
+    /// Creates a network request that retrieves the contents of a URL \
+    /// based on the specified URL request object and returns the data on completion
     ///
-    /// Start a network request
     ///
-    /// - parameter url: The url to be parsed
-    /// - parameter posting: What do you need to post
-    /// - returns: Result<String, Error>
-    @available(*, deprecated)
-    public func networkRequest(
-        url: String,
-        posting: [String: Any]? = ["nothing": "send"]
-    ) -> Result<String, Error> {
-        // Ok, this is a waiter
-        var rResult: Result<String, Error>?
-        
-        self.networkRequest(url: url, posting: posting) { (res) in
-            rResult = res
-        }
-        
-        while rResult == nil {
-            // wait.
-        }
-        
-        return rResult ?? .failure(
-            AuroraError(message: "Failed to unwrap result")
-        )
-    }
-    
-    /// NOT DONE YET (Blocking)
+    ///
     /// - Parameters:
-    ///   - url: <#url description#>
-    ///   - method: <#method description#>
-    ///   - values: <#values description#>
-    /// - Returns: <#description#>
+    ///   - url: A value that identifies the location of a resource, \
+    ///   such as an item on a remote server or the path to a local file.
+    ///   - method: The HTTP request method.
+    ///   - values: POST values (if any)
+    /// - Returns: The HTTP(S) request data
     public func networkRequest(
         url: String,
         method: HTTPMethod,
@@ -165,8 +120,9 @@ extension Aurora {
         var result: Data?
         var inGroup = true
         var waiting = true
+        let dGroup = Aurora.group
         
-        Aurora.group.enter()
+        dGroup.enter()
         
         DispatchQueue.main.asyncAfter(deadline: .now() + 30) {
             if waiting {
@@ -175,7 +131,7 @@ extension Aurora {
                 )
                 
                 if inGroup {
-                    Aurora.group.leave()
+                    dGroup.leave()
                 }
             }
         }
@@ -188,34 +144,40 @@ extension Aurora {
             
             switch response {
             case .success(let data):
-                result = data.data(using: .utf8)
+                result = data
             case .failure(let error):
                 Aurora.shared.log(error.localizedDescription)
             }
             
             if inGroup {
-                String.group.leave()
+                dGroup.leave()
             }
         }
         
-        String.group.wait()
-        
+        dGroup.wait()
+                
         inGroup = false
         return result
     }
     
     // swiftlint:disable function_body_length
-    /// [Unfinished]
+    /// Creates a network request that retrieves the contents of a URL \
+    /// based on the specified URL request object, and calls a handler upon completion.
+    ///  
     /// - Parameters:
-    ///   - url: <#url description#>
-    ///   - method: <#method description#>
-    ///   - values: <#values description#>
-    ///   - completionHandler: <#completionHandler description#>
+    ///   - url: A value that identifies the location of a resource, \
+    ///   such as an item on a remote server or the path to a local file.
+    ///   - method: The HTTP request method.
+    ///   - values: POST values (if any)
+    ///   - completionHandler: This completion handler takes the following parameters:
+    ///   `Result<Data, Error>`
+    ///     - `Data`: The data returned by the server.
+    ///     - `Errror`: An error object that indicates why the request failed, or nil if the request was successful.
     public func networkRequest(
         url: String,
         method: HTTPMethod,
         values: [String: Any]?,
-        completionHandler: @escaping (Result<String, Error>) -> Void
+        completionHandler: @escaping (Result<Data, Error>) -> Void
     ) {
         /// Check if the URL is valid
         guard let siteURL = URL(string: url) else {
@@ -233,19 +195,21 @@ extension Aurora {
         
         // Try
         do {
-            /// Create JSON
-            let JSON = try JSONSerialization.data(
-                withJSONObject: values as Any,
-                options: .sortedKeys
-            )
-            
-            // set NewPosting
-            post = String.init(
-                data: JSON,
-                encoding: .utf8
-            )!.addingPercentEncoding(
-                withAllowedCharacters: .urlHostAllowed
-            )!
+            if let safeValues = values {
+                /// Create JSON
+                let JSON = try JSONSerialization.data(
+                    withJSONObject: safeValues,
+                    options: .sortedKeys
+                )
+                
+                // set NewPosting
+                post = String.init(
+                    data: JSON,
+                    encoding: .utf8
+                )!.addingPercentEncoding(
+                    withAllowedCharacters: .urlHostAllowed
+                )!
+            }
         }
         
         /// Catch errors
@@ -294,7 +258,7 @@ extension Aurora {
         }
         
         /// Create a pinned URLSession
-        var session = URLSession.init(
+        var session: URLSession? = URLSession.init(
             // With default configuration
             configuration: .default,
             
@@ -317,7 +281,7 @@ extension Aurora {
         }
         
         if let cookieData = Aurora.cookies {
-            session.configuration.httpCookieStorage?.setCookies(
+            session?.configuration.httpCookieStorage?.setCookies(
                 cookieData,
                 for: siteURL,
                 mainDocumentURL: nil
@@ -325,7 +289,7 @@ extension Aurora {
         }
         
         // Start our datatask
-        session.dataTask(with: request) { (sitedata, _, theError) in
+        session?.dataTask(with: request) { (sitedata, _, theError) in
             /// Check if we got any useable site data
             guard let sitedata = sitedata else {
                 if post.length > 3 {
@@ -339,7 +303,7 @@ extension Aurora {
             }
             
             // Save our cookies
-            Aurora.cookies = session.configuration.httpCookieStorage?.cookies
+            Aurora.cookies = session?.configuration.httpCookieStorage?.cookies
             
             if post.length > 3 {
                 let data = String.init(data: sitedata, encoding: .utf8)
@@ -352,166 +316,18 @@ extension Aurora {
             }
             
             completionHandler(
-                .success(
-                    String.init(data: sitedata, encoding: .utf8)!
-                )
+                .success(sitedata)
             )
         }.resume()
+        
+        // Invalidate session, after saving tasks
+        session?.finishTasksAndInvalidate()
+        
+        // Release the session from memory
+        session = nil
     }
     // swiftlint:enable function_body_length
     
-    // swiftlint:disable function_body_length
-    /// networkRequest (non-blocking)
-    ///
-    /// Start a network request
-    ///
-    /// - parameter url: The url to be parsed
-    /// - parameter posting: What do you need to post
-    /// - returns: closure -> sucess, fail.
-    public func networkRequest(
-        url: String,
-        posting: [String: Any]?,
-        completionHandler: @escaping (Result<String, Error>) -> Void
-    ) {
-        
-        /// Check if the URL is valid
-        guard let siteURL = URL(string: url) else {
-            completionHandler(
-                .failure(
-                    AuroraError(message: "Error: \(url) doesn't appear to be an URL")
-                )
-            )
-            
-            return
-        }
-        
-        /// Create a new post dict, for the JSON String
-        var post: String = ""
-        
-        if posting != nil {
-            // Try
-            do {
-                /// Create JSON
-                let JSON = try JSONSerialization.data(
-                    withJSONObject: posting as Any,
-                    options: .sortedKeys
-                )
-                
-                // set NewPosting
-                post = String.init(
-                    data: JSON,
-                    encoding: .utf8
-                )!.addingPercentEncoding(
-                    withAllowedCharacters: .urlHostAllowed
-                )!
-            }
-            
-            /// Catch errors
-            catch let error as NSError {
-                completionHandler(
-                    .failure(
-                        AuroraError(message: "Error: \(error.localizedDescription)")
-                    )
-                )
-            }
-        }
-        
-        /// Create a URL Request
-        let request = URLRequest(url: siteURL).configure {
-            // 60 Seconds before timeout (default)
-            $0.timeoutInterval = 15
-            // Set Content-Type to FORM
-            $0.setValue("close", forHTTPHeaderField: "Connection")
-            $0.setValue(self.userAgent, forHTTPHeaderField: "User-Agent")
-            
-            if post.length > 3 {
-                // We're posting
-                
-                // Set the HTTP Method to POST
-                $0.httpMethod = "POST"
-                
-                // Set Content-Type to FORM
-                $0.setValue(
-                    "application/x-www-form-urlencoded",
-                    forHTTPHeaderField: "Content-Type"
-                )
-                
-                // Set the httpBody
-                $0.httpBody = post.data(using: .utf8)
-            } else {
-                // We're getting
-                
-                // Set the HTTP Method to GET
-                $0.httpMethod = "GET"
-            }
-        }
-        
-        /// Create a pinned URLSession
-        var session = URLSession.init(
-            // With default configuration
-            configuration: .default,
-            
-            // With our pinning delegate
-            delegate: AuroraURLSessionPinningDelegate(),
-            
-            // with no queue
-            delegateQueue: nil
-        )
-        
-        // Check if we have a public key, or certificate hash.
-        if HTTPSCertificate.publicKeyHash.count == 0 || HTTPSCertificate.certificateHash.count == 0 {
-            // Show a error, only on debug builds
-            log(
-                "[WARNING] No Public key pinning/Certificate pinning\n" +
-                    "           Improve your security to enable this!\n"
-            )
-            // Use a non-pinned URLSession
-            session = URLSession.shared
-        }
-        
-        if let cookieData = Aurora.cookies {
-            session.configuration.httpCookieStorage?.setCookies(
-                cookieData,
-                for: siteURL,
-                mainDocumentURL: nil
-            )
-        }
-        
-        // Start our datatask
-        session.dataTask(with: request) { (sitedata, _, theError) in
-            /// Check if we got any useable site data
-            guard let sitedata = sitedata else {
-                if post.length > 3 {
-                    self.log("Error: \(theError?.localizedDescription)")
-                } else {
-                    self.log("Error: \(theError?.localizedDescription)")
-                }
-                
-                completionHandler(.failure(theError!))
-                return
-            }
-            
-            // Save our cookies
-            Aurora.cookies = session.configuration.httpCookieStorage?.cookies
-            
-            if post.length > 3 {
-                let data = String.init(data: sitedata, encoding: .utf8)
-                
-                Aurora.fullResponse = data
-            } else {
-                let data = String.init(data: sitedata, encoding: .utf8)
-                
-                Aurora.fullResponse = data
-            }
-            
-            completionHandler(
-                .success(
-                    String.init(data: sitedata, encoding: .utf8)!
-                )
-            )
-        }.resume()
-    }
-    // swiftlint:enable function_body_length
     
     /// Return the full networkRequestResponse
     /// - Returns: the full networkRequestResponse
