+++ Sources/Aurora/Classes/AuroraLogview.swift
@@ -9,15 +9,17 @@
 import Foundation
 import UIKit
 
-/// <#Description#>
+/// Aurora logviewer class
+///
+/// View the logs used/generated by Aurora.Framework
 public class AuroraLogView: UIViewController, UITableViewDelegate, UITableViewDataSource {
-    /// <#Description#>
+    /// Log messages
     var logMessages = Aurora.shared.getLogMessages()
     
-    /// <#Description#>
+    /// The tableview
     var tableView: UITableView = UITableView()
     
-    /// <#Description#>
+    /// View did load
     public override func viewDidLoad() {
         logMessages.append("END OF REPORT\r\n\r\n")
         logMessages.append("AuroraExitLogViewButton")
@@ -44,7 +46,7 @@ public class AuroraLogView: UIViewController, UITableViewDelegate, UITableViewDa
         }
     }
     
-    /// <#Description#>
+    /// reloads the logview
     private func reloadLogView() {
         logMessages = Aurora.shared.getLogMessages()
         logMessages.append("END OF REPORT\r\n\r\n")
